// src/models/types.ts
export enum ECodeRuleType {
  Literal = "literal",
  ParameterValue = "parameterValue",
  Year = "year",
  Month = "month",
  DayOfMonth = "dayOfMonth",
  DayOfWeek = "dayOfWeek",
  DayOfYear = "dayOfYear",
  WeekOfYear = "weekOfYear",
  AutoIncrement = "autoIncrement",
  Checksum = "checksum",
  UserInput = "userInput",
}

export enum ELabelTextAlign {
  Left = "left",
  Right = "right",
  Center = "center",
}